+++ lib/shelly/download_progress_bar.rb
@@ -1,14 +1,16 @@
-require "progressbar"
+require "ruby-progressbar"
 
 module Shelly
-  class DownloadProgressBar < ProgressBar
-    def initialize(total)
-      super("Progress", total)
-      self.format_arguments = [:title, :percentage, :bar, :stat_for_file_transfer]
+  class DownloadProgressBar < ProgressBar::Base
+    def initialize(total = nil)
+      super(:title => "Progress", :total => total, :format => "%a [%B] %p%% %t | %E")
     end
 
     def progress_callback
-      lambda { |size| inc(size) }
+      lambda { |inc_size, total_size|
+        self.total = total_size if total_size
+        self.progress = progress + inc_size
+      }
     end
   end
 end